requirements.txt
Django~=2.2.4
mysqlclient==2.0.3
djangorestframework==3.12.4